==> 00_English_LLAMA.ipynb <==
# Introduction to the problem

import os
import torch
from datasets import load_dataset
from transformers import (
    AutoModelForCausalLM,
    AutoTokenizer,
    BitsAndBytesConfig,
    TrainingArguments,
    pipeline,
    logging,
)
from peft import LoraConfig, PeftModel
import time

## Introduction

In this notebook, I want to explore the power of the LLAMA 2 model which is an open source model LLM created by Meta and given a fine-tunned version trained on instructions. Since I have trouble picking a meal for the day due to my busy schedule, I want to use this model to give me fine Bulgarian dishes. Even more, I want to read and use the model in Bulgarian language. There are many people (older generations) in Bulgaria who are not able to read and write in English and this work will investigate how well can this model be thought to generate Bulgarian recipes in Bulgarian language.

The powerful side of LLAMA models is that they can be trained with your private data and you can host the model on your own infrastructure with available GPU on it. The technique for fine-tuning this model is called LORA and with the help of Quantization we can reduce the size of the wights on our GPU so we can save some space but the precision is lost in the process.

Infrastructure info:

AWS EC2 instance - g5.xlarge
                 - AMI ami-09adc81b646216f66

## Load Model

english_model = "NousResearch/Llama-2-7b-chat-hf"

Here I am using the BitsAndBytes library to reduce the model size.

compute_dtype = getattr(torch, "float16")

quant_config = BitsAndBytesConfig(
    load_in_4bit=True,
    bnb_4bit_quant_type="nf4",
    bnb_4bit_compute_dtype=compute_dtype,
    bnb_4bit_use_double_quant=False,
)

model = AutoModelForCausalLM.from_pretrained(
    english_model,
    quantization_config=quant_config,
    device_map={"": 0}
)
model.config.use_cache = False
model.config.pretraining_tp = 1

tokenizer = AutoTokenizer.from_pretrained(english_model)
tokenizer.pad_token = tokenizer.eos_token
tokenizer.padding_side = "right"

!nvidia-smi

We can see after we load the original model that it needs nearly 5GB of GPU momery for a model having 7B parameters and qunatized at 4bit version. Using the 8bit version will make 13GB and to use the original 16bit model will take all the way to 22GB.

def generate(text, max_length=500, temperature=0.3):
    # Ignore warnings
    start_time = time.time()
    logging.set_verbosity(logging.CRITICAL)
    pipe = pipeline(task="text-generation", model=model, tokenizer=tokenizer, max_length=max_length, temperature=temperature)

    # Run text generation pipeline with our next model
    prompt = text
    result = pipe(f"[INST] {prompt} [/INST]")
    end_time = time.time()
    execution_time = end_time - start_time
    print("Execution time:",execution_time)
    print(result[0]['generated_text'])

generate("What kind of LLM are you?")

generate("Do you know recipes for dishes?")

generate("Do you know recipes for Bulgarian dishes? If yes give me an example")

generate("Do you know to speak bulgarian?")

|  English | Generated  | Real Bulgarian  |
|---|---|---|
| Goodbye   | До свидание  | Довиждане  |
|  I'm fine, thank you | Имам добро, бяло  | Добре съм, благодаря ти  |
|  My name is |  Имам името | Името ми е  |
|  Please  | Пожалуйста   |  Моля |

# What is AI?
generate("Какво е Изкуствен интелект?")

Errors from the text:
 - Кünstльный - its a mixture of german and russian
 - It says there are 2 types of artiffical intelligence: 
    1. Natural intelligence - Doesn't make any sense
- животта is with one **T**


# Tell me how to make tarator (Bulgarian dish)
generate("Кажи ми как се готви таратор?")

Firstly the generated text is in English and also the recipe is incorrect. Tarator is a dish with yogurt cucumbers, dill, and salt. Perfect for drinking in the summer.

# Tell me what to make today? I have potatoes, meat and spices
generate("Какво да си сготвя днес? Имам картофи, месо. подправки.")

As we see, the generation is in egnlish and even more, I want to be given jsut one recepie! I don't have the time to read all of them and also what are the steps of making Potato and meet tacos? Are there any potatos in tacos in a first place?
