==> 01_Data_prep.ipynb <==
# Data Preparation

import pandas as pd
import re
import random
import numpy as np
from sklearn.model_selection import train_test_split

df = pd.read_table("./data/bg-recipes.tsv")
df.drop("Unnamed: 0", axis=1, inplace=True)

df

Our data consists of the title of the dish, rating from the website, products with the quantities, and the description of how to make it.

txt = df["products"][0]
txt

# Remove [EOL] and other non-alphabetic characters

def clean_txt(txt):
    text_with_commas = re.sub(r'\[EOL\]', ', ', txt)
    cleaned_text = re.sub(r'ч\.л\.|\d|[-./]|ч\.ч\.|бр\.|с\.л\.|\bг\b|\bкг\b|\bмл\b|\bч л\b|\b( с л )\b', ' ', text_with_commas)
    # Replace multiple spaces with a single space
    cleaned_text = re.sub(r'\s+', ' ', cleaned_text).strip()

    return cleaned_text

We have to clean the text a bit with removing all the [EOL], and all product quantities like (soup spoon, cup, kg, ml and etc.)

random.seed(10)

randomlist = []
for i in range(0,50):
    n = random.randint(1,df.shape[0])
    randomlist.append(n)
    picked_txt = df["products"][n]
    print(picked_txt)
    print(clean_txt(picked_txt))
    print("\n")
    print("--------------------------------------")

Now we have a better ingredients column consisting only of them and separated by a comma.

df["cleaned_products"] = df["products"].map(clean_txt)
df

text_df1 = pd.DataFrame()
text_df2 = pd.DataFrame()

For fine-tunning the model we need to follow the LLAMA standard provided by Meta itself. The trianing data should consist of the following text structure:

[INST] Instruction that we are going to give [/INST] The answer we are expecting from the model to be given.


text_df1["promped_text"] = "<s>[INST] Искам да сготвя нещо хубаво. Имам налични следните продукти: " + df["cleaned_products"] + " [/INST]" + " Може да приготвиш " + df["title"] + ": " + df["description"]

My model input will look like that:
    [INST] I want to cook somehting nice. I have the follwoing products: product1, product2, peoduct3 [/INST] You can cook `Title of dish: Dish description.`

text_df1.promped_text[0]

[INST] I want to cook something nice. I have the following products available: boiled beans, pickles, sauerkraut, potatoes, oil, vinegar, onions [/INST] You can prepare a winter salad: Mash the beans with a fork, and finely chop the boiled potatoes, cucumbers and cabbage, onion is cut into slices. Everything is mixed very well and seasoned.


text_df2["promped_text"] = "<s>[INST] Гладен съм. Какво мога да си сготвя със: " + df["cleaned_products"] + " [/INST]" + " Начин на приготвяне на " + df["title"] + ": " + df["description"]

I created also another sentence starter: I'm hungry. What can I cook with:

I duplicate the data in a way to achieve a better understanding of the language and expectation of new ways to provide a command.

result_df = pd.concat([text_df1, text_df2], ignore_index=True)

result_df

train_data, test_data = train_test_split(result_df, shuffle=True, train_size=0.85)

train_data.to_csv("./data/train_data.csv", index=False)
test_data.to_csv("./data/test_data.csv", index=False)
