==> 02_Trianing.ipynb <==
# Trianing and saving the model

We can see with load model we have 5GB out of 23 with 4bit quantized model. 

!nvidia-smi

!python --version

!nvcc --version

## Issues with the project I saw

Since I don't have such a powerful GPU on my private laptop I am using AWS infrastructure in that case. One of the problems is that the AWS EC2 instance is using old AMI images (pre-defined Linux snapshots with installed libraries). Using this AMI we are using the old version of CUDA which means using the torch==1.13 version and updating the CUDA is a challenge itself! That's why I needed to install the respective versions of Pytorch and transformers. 
The next command is to uninstall one problematic **nvidia_cublas_cu11** driver.

#!yes | pip uninstall nvidia_cublas_cu11

import os
import torch
from datasets import load_dataset
from transformers import (
    AutoModelForCausalLM,
    AutoTokenizer,
    BitsAndBytesConfig,
    TrainingArguments,
    pipeline,
    logging,
)
from peft import LoraConfig
from trl import SFTTrainer
import time

hf_link = "NousResearch/Llama-2-7b-chat-hf"
data = "./data/train_data.csv"
new_model = "./model/new_model/"
base_model = "./model/base_model/"

train_data = load_dataset("csv", data_files=data)
train_data = train_data.rename_column("promped_text", "text")

train_data

train_data["train"][0]

Creating the configuration for 4bit model quantization.

compute_dtype = getattr(torch, "float16")

quant_config = BitsAndBytesConfig(
    load_in_4bit=True,
    bnb_4bit_quant_type="nf4",
    bnb_4bit_compute_dtype=compute_dtype,
    bnb_4bit_use_double_quant=False,
)

For using the model and training, it is important to have GPU and CUDA.

# Check GPU compatibility with bfloat16
if compute_dtype == torch.float16:
    major, _ = torch.cuda.get_device_capability()
    if major >= 8:
        print("=" * 80)
        print("Your GPU supports bfloat16: accelerate training with bf16=True")
        print("=" * 80)

Loading the model and the tokenizer in GPU.
The model is being donwloaded from the Huggingface repo and its stored in the following path (Linux): ~/.cache/huggingface/hub/name_of_the_hf_repo/snapshots/random_id/

model = AutoModelForCausalLM.from_pretrained(
    hf_link,
    quantization_config=quant_config,
    device_map={"": 0}
)
model.config.use_cache = False
model.config.pretraining_tp = 1

tokenizer = AutoTokenizer.from_pretrained(hf_link)
tokenizer.pad_token = tokenizer.eos_token
tokenizer.padding_side = "right"

#!sudo cp -L -R ~/.cache/huggingface/hub/models--NousResearch--Llama-2-7b-chat-hf/snapshots/37892f30c23786c0d5367d80481fa0d9fba93cf8/* ./model/base_model

Lora Config is responsible for setting up the trainable parameters for the adapter model. This model is responsible for gaining knowledge for the fine-tunning part. The values are pretty standard based on the documentation. 

peft_params = LoraConfig(
    lora_alpha=16,
    lora_dropout=0.1,
    r=64,
    bias="none",
    task_type="CAUSAL_LM",
)

Setting up the training parameters and initializing the trainer.

It is recommended to use 1 epoch for training since it overfits fast. Also with the given infrastructure capacity, I am using a maximum 5 data instances per step to go for training. Logging steps are 100 and the learning rate is the recommended one: 2e-4.


training_params = TrainingArguments(
    output_dir="./results",
    num_train_epochs=1,
    per_device_train_batch_size=5,
    gradient_accumulation_steps=1,
    optim="paged_adamw_32bit",
    save_steps=2000,
    logging_steps=100,
    learning_rate=2e-4,
    weight_decay=0.001,
    max_grad_norm=0.3,
    max_steps=-1,
    warmup_ratio=0.03,
    group_by_length=True,
    lr_scheduler_type="constant"
)

trainer = SFTTrainer(
    model=model,
    train_dataset=train_data["train"],
    peft_config=peft_params,
    dataset_text_field="text",
    max_seq_length=None,
    tokenizer=tokenizer,
    args=training_params
)

# Train model
trainer.train()

After more than 6 and a half hours of training time, finaly the model is being saved and ready to use.

trainer.model.save_pretrained(new_model)
trainer.tokenizer.save_pretrained(new_model)

# Ignore warnings
logging.set_verbosity(logging.CRITICAL)

# Run text generation pipeline with our next model
prompt = "What is a large language model?"
pipe = pipeline(task="text-generation", model=model, tokenizer=tokenizer, max_length=200)
result = pipe(f"[INST] {prompt} [/INST]")
print(result[0]['generated_text'])

pipe = pipeline(task="text-generation", model=model, tokenizer=tokenizer, max_length=200)

def generate(text, pipeline):
  # Ignore warnings

  start_time = time.time()
  logging.set_verbosity(logging.CRITICAL)

  # Run text generation pipeline with our next model
  prompt = text
  result = pipeline(f"[INST] {prompt} [/INST]")
  end_time = time.time()
  execution_time = end_time - start_time
  print("Execution time:",execution_time)
  print(result[0]['generated_text'])
  return result[0]['generated_text']

Inspect some generations:

generate("Какво да си сготвя днес? Имам картофи, месо. подправки.", pipe)

[INST] What should I cook today? I have potatoes, meat. spices. [/INST] How to prepare Potatoes with meat: Potatoes are boiled, peeled and cut into slices. The meat is boiled, cut into pieces and mixed with the potatoes. Season to taste. If desired, cheese can be added. 10 minutes in the microwave. 10 minutes in the microwave. 10 minutes in the microwave. 10 minutes in the microwave. 10 minutes in the microwave. 1

generate("Можеш ли да говорш на български?", pipe)

[INST] Can you speak Bulgarian? [/INST] The voiced way of speaking Bulgarian is better because all the sounds are tried and distributed better. everybody can speak Bulgarian in a fun and easy way. 1. Start with the vowels. 2. Learn the sounds of the letters. 3. Practice the sounds of the letters. 4. Learn the words. 5. Practice the words. 6. Learn the grammar. 7. Practice the grammar. 8. Learn the idioms. 9. Practice the idioms. 10. Learn the songs. 11. Practice the songs. 12. Learn the poems. 13. Practice the po

generate("Имам ориз, мляко. какъв десерт да наравя?", pipe)

[INST] I have rice, milk. what dessert should i make? [/INST] You can prepare Rice with milk: Rice is boiled in salted water. Drain and place in a bowl. The milk is heated and poured over the rice. Mix well. Cinnamon can be added if desired. If desired, you can also add a banana. Leave to harden for 10-15 minutes. Served with whipped cream if desired. 10-15 minutes before serving, sprinkle with cinnamon. 10-15 minutes before serving
