==> 03_Load_using_model.ipynb <==
# Testing the new model

import os
import torch
from datasets import load_dataset
from transformers import (
    AutoModelForCausalLM,
    AutoTokenizer,
    BitsAndBytesConfig,
    TrainingArguments,
    pipeline,
    logging,
)
from peft import LoraConfig, PeftModel, PeftModelForCausalLM
import time
import pandas as pd
import random
from nltk.translate.bleu_score import SmoothingFunction, corpus_bleu, sentence_bleu

## Load the model

new_model = "teodor98/LORA_Config_BG_Recepies"
base_model = "NousResearch/Llama-2-7b-chat-hf"


After the training I created a repository on Huggingface where I stored my adapter model.
You can check the following link: [Huggingface repo](https://huggingface.co/teodor98/LORA_Config_BG_Recepies)

compute_dtype = getattr(torch, "float16")

quant_config = BitsAndBytesConfig(
    load_in_8bit=True,
    bnb_8bit_quant_type="nf4",
    bnb_8bit_compute_dtype=compute_dtype,
    bnb_8bit_use_double_quant=False,
)

model = AutoModelForCausalLM.from_pretrained(
    base_model,
    #quantization_config=quant_config,
    low_cpu_mem_usage=True,
    return_dict=True,
    torch_dtype=torch.float16,
    device_map={"": 0}
)
model.config.use_cache = False
model.config.pretraining_tp = 1

I face a problem with loading the model and the adapter. When I load the base model Llama 7B via quantization I cannot use PeftModel merge_and_unload. I can't merge LORA layers with an already quantized model.

For now, I am loading the whole Llama model and then I am merging it with the fine-tuned adapter.

bg_model = PeftModel.from_pretrained(model, new_model)
bg_model = bg_model.merge_and_unload()

!nvidia-smi

Because of the problem I my GPU memory grows up to 14GB. Which is a problem if I dont have the optimal infrastructure.

I tried also putting the english model and my adapter in one folder but still it loads only the english base model.

tokenizer = AutoTokenizer.from_pretrained(new_model)
tokenizer.pad_token = tokenizer.eos_token
tokenizer.padding_side = "right"

def generate(text, max_len=200):
  # Ignore warnings
  pipe = pipeline(task="text-generation", model=bg_model, tokenizer=tokenizer, max_length=max_len)
  start_time = time.time()
  logging.set_verbosity(logging.CRITICAL)

  # Run text generation pipeline with our next model
  prompt = text
  result = pipe(f"[INST] {prompt} [/INST]")
  end_time = time.time()
  execution_time = end_time - start_time
  print("Execution time:",execution_time)
  #print(result[0]['generated_text'])

  return result[0]['generated_text']

_ = generate("Какво да си сготвя днес? Имам картофи, месо. подправки.")
_ 

_ = generate("Знаеш ли как мога да тренирам невронна мрежа?")
_ 

This general knowleadge answer is far from correct but we can see that is sounds like recepie (slap in bulgarian means to  cover the pot):

'[INST] Do you know how I can train a neural network? [/INST] Anyone damaged by neural nets can blend in with them. *clap* 10 minutes and *clap* 10 minutes. 10 minutes of *slap* and *slap*. 10 minutes of *slap* and *slap*. 10 minutes of *slap* and *slap*. 10 minutes of *slap* and *slap*. 10 minutes of *slap* and *slap*. 10 minutes of *slap* and *slap*. 10 minutes'

test_data = pd.read_csv("./data/test_data.csv")

test_data

# Splitting the 'prompted_text' column
test_data[['instruction', 'response']] = test_data['promped_text'].str.split(r'\s*\[/INST\]\s* ', expand=True)

test_data.instruction = test_data.instruction.str.replace("<s>[INST] ", "")
test_data

random.seed(42)

random_numbers = random.sample(range(test_data.shape[0]), 10)

for i in random_numbers:
    picked_txt = test_data.instruction[i]
    print("Instruction:")
    print(picked_txt)
    print("The Real answer is: ")
    print(test_data.response[i] + "\n")
    print("The Generated answer is:")

    gen_txt = generate(picked_txt, max_len=500)
    print(gen_txt.split("[/INST] ")[1])

    print("================================================= \n")

You can check out the **English_generations.docx** document which has also the translations of the text. 

As we can see the generations are surprisingly good. Not only I am getting Bulgarian-generated and grammatically correct sentences, but I also get pretty diverse recipes. Sometimes I get a different dish than the one in the test set but also makes sense with all the ingredients. That indicates that the model doesn't overfit as much to learn by hard the answers.
As problem is the repetitiveness of end sentences from the generation. 

generate("Имам шоколад и люти чушки, какво да сготвя?")

I asked him to generate a recepie with chocolate and spicy peppers and it generated this:
You can prepare Chocolate with hot peppers: Put the chocolate in a bowl and put it in a water bath until it melts. 1-2 peppers are cut into pieces and put in the chocolate. Cinnamon can be added if desired. Let the chocolate and peppers melt. 2-3 peppers are cut into pieces and put in the chocolate. Let the chocolate and peppers melt. Cinnamon can be added if desired. 4-5 peppers
It generates the same sentence all over again but make the peppers quantity greater and greater. It sounds confident at its recepie


generate("кажи ми как титаник е потънал?")

I asked him to tell me how Titanic sinked and the answer is following:

What happened to the Titanic is well known.
The Titanic was a passenger cruise liner that sank on April 15, 1912 in the Atlantic Ocean. The sunken Titanic is the largest and most well-known of all sunken ships in history. The Titanic Wreck is a passenger cruise liner, 234m long and 27m wide. The Titanic Wreck has a capacity of 2,640 passengers and 800'

I dont get the concrete answer but just explanation that I dont need that much, but at least it generates gramatically correct sentances.

generate("генереирай код на java който минава през цикъл със имена на хора и принтира в конзолата: Аз се казвам името на човека, как се казваш ти?")

Here we can see that the model does not generate any kind of Java code:

'[INST] generate java code that loops through people's names and prints to the console: My name is person's name, what's your name? [/INST] Can add all people's names to the console and try to fight them all.\n\n1. My name is the person's name, what is your name?\n2. My name is the person's name, what is your name?\n3. My name is the person's name, what is your name?\n4. My name is the person's name, what is your name?\n5. I'm going

Lets try asking in english

generate("generate java code that loops through people's names and prints to the console: My name is person's name, what's your name?")

'[INST] generate java code that loops through people\'s names and prints to the console: My name is person\'s name, what\'s your name? [/INST]  Sure! Here is an example of how you can loop through people\'s names and print them to the console in Java: 
```import java.util.Scanner;
public class NameGenerator {
      public static void main(String[] args) {
           Scanner scanner = new Scanner(System.in);
            while (scanner.hasNext()) {
                     String name = scanner.next();
                     System.out.println("My name is " + name);
                     System.out.println("What\'s your name?"); }
                      }
                }\```
                
This code uses a `Scanner` object to read input from the user, and a `while` loop to repeat the process until the user enters nothing.'

It is not that accurate but still in english it understands and does way more.

In conclusion, I can say that:
- The model can be fine-tuned on a task-specific job and do a pretty good job at it (at least with a big data set like ours)
- My task was to get Bulgarian grammatically correct sentences but it taught another task as well: Generating recipes!
- Still, the model is not good for general knowledge questions.
- It needs huge GPU capacity because does not allow to quantize the model and the fine-tunned adaptor at the same time.
